requirements.txt
streamlit==1.32.0
numpy==1.24.3
opencv-python-headless==4.9.0.80
Pillow==10.2.0
matplotlib==3.8.3
nashpy==0.0.40
scipy==1.12.0
fpdf2==2.7.8
